requirements.txt
# Core deep learning stack (match README instructions; install torch/torchaudio with the CUDA 12.4 wheels)
torch>=2.5.1,<=2.8.0
torchaudio>=2.5.1,<=2.8.0

# General utilities
numpy
tqdm
fire
einops
omegaconf==2.3.0
torchmetrics

# Text and tokenizer support
sentencepiece
transformers==4.57.3
huggingface_hub
datasets<4
peft>=0.12.0

# Audio/tokenizer backends
openai-whisper
https://huggingface.co/NandemoGHS/Anime-XCodec2-44.1kHz-v2/resolve/main/xcodec2-0.1.7.tar.gz
ffmpeg-python
scipy
soundfile
librosa

# Training / logging
tensorboard
wandb
matplotlib

# Inference / demos
gradio
torchcodec

# Duration estimation (phoneme/lang aware)
langdetect
g2p_en
pyopenjtalk-plus[onnxruntime]
pypinyin
